Fixed issue when -c <client> was specified and there are Migrated jobs
When -c <client> is specified, all Migration Ctrl and Copy Ctrl jobs are omitted
from the initial db query. This being the case, iot there are jobs in the list that
have been Migrated (jobstatus='M'), there will be no Migration Ctrl job to make
the connection to the job it was migrated to. In this case, now just print "Migrated"