Now includes copied and migrated jobs from outside the specified "-t hour" interval
Copied an migrated jobs inherit the endtime of the original backup job. In many cases
Copy Ctrl and Migrtion Ctrl jobs in the listing would reference jobids that were not
in the list. This would be confusing and would force the user to have to manually look
up the original job in bconsole, BWeb, or Baculum.

Now, backup jobs, or copies, or migrated jobs will be included in the list of jobs, even
if their endtime is outside of the "-t hours" set.