Bug fix for "NameError: name 'pn_jobids_lst' is not defined`" reported by 'niccarp'
Error was due to that variable only being defined if/when there are previous/new jobs.

With no copy/migration jobs in the "-t hours" listing, this variable is not created,
yet I had incorrectly referenced it later to add a banner at the top of the email, and
to prepend job's endtimes with an asterisk.