Fixed a bug about not counting files/bytes of Copied/Migrated jobs in the Summary
This bug was due to a decision I made a long time ago to be able to pull "old" jobs (outside of -t hours)
into the jobs list that had been copied or migrated, but not count their files/bytes.

Previously, I was counting files/bytes for 'M'igrated 'C'opied jobs that were within the -t hours. The
modification here was to count the files/bytes of the Copy Control (c) and Migration Control (g) jobs
instead. My logic here is that if a Copy/Migration job runs within the -t hours, then the data it moved
should be counted in the final Copied/Migrated stats, even if the job it worked on was not.

Of course, the files/bytes for these pulled in "old" backup jobs will not be counted against the other
statisitics. ie: Total jobs, Bad Jobs, Total Job Errors, Total Backup Files/Bytes