Workaround for Copy/Migration Control jobs not inserting File/Bytes into catalog
This is to solve the issue where versions of Bacula
Community < 13.0 and Bacula Enterprise < 14.0 did not put
the jobfiles and jobbytes of Copy/Migrate Control jobs into
the catalog. The only other place to find this information
is in the Job's Summary text blob in the 'SD Files Written:'
and 'SD Bytes Written:' lines, so this workaround now takes
this information from there and will work with all versions
of Bacula, Community or Enterprise.