Added two screenshots showing the new Antivirus feature
Modified the Usage section to include the new '-a <avemail>' command line parameter which allows Virus Report emails to be sent to a different address than the jobs report. If not specified, the default is to send the Virus Report email to the same email address as the jobs report.